Move RocketConfigs to rocket.Configs
